Abstracted out paidfor and supported code into shared js file.
