+++ src/capi-multiple-paidfor/index.html
@@ -8,7 +8,7 @@
 					</button>
 					<div class="popup popup--paidfor is-off js-paidfor-popup">
 						<h3 class="popup--paidfor__title">Paid content is paid for and controlled by an advertiser and produced by the Guardian Labs team</h3>
-						<a class="popup--paidfor__link" href="https://theguardian.com/content-funding">Learn more about Guardian Labs content <span class="inline-arrow-down inline-icon">{{#svg}}arrow-right{{/svg}}</span>
+						<a class="popup--paidfor__link" href="%%CLICK_URL_UNESC%%https://theguardian.com/content-funding">Learn more about Guardian Labs content <span class="inline-arrow-down inline-icon">{{#svg}}arrow-right{{/svg}}</span>
 						</a>
 					</div>
 				</div>
