+++ src/capi-multiple-supported/index.html
@@ -8,7 +8,7 @@
 			<a class="badge__link" href="%%CLICK_URL_UNESC%%[%SeriesURL%]" data-link-name="logo link">
 				<img class="badge__logo" src="[%BrandLogo%]" alt="">
 			</a>
-			<a href="%%CLICK_URL_UNESC%%https://theguardian.com/guardian-labs" class="badge__help" data-link-name="about link">About this content</a>
+			<a href="%%CLICK_URL_UNESC%%https://theguardian.com/about" class="badge__help" data-link-name="about link">About this content</a>
 		</div>
 	</header>
 	<div class="adverts__body">
